requirements.txt
aiogram>=3.3.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz
aiohttp>=3.9.0
psycopg2-binary>=2.9.0
